# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Bandage API
  version: "0.1.0"
  description: >
    Contract for the Smart Bandage backend (Phase 4). Drafted before the
    backend exists so the dashboard, mobile app, and simulator can all be
    written against it. Simulator-produced and hardware-produced
    measurements enter through the exact same endpoints.
servers:
  - url: http://localhost:8000
paths:
...
  /devices/register:
    post:
      summary: Register a device (physical or simulated)
      tags: [devices]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Device" }
      responses:
        "201":
          description: Device registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Device" }
...
  /devices/{device_id}:
    get:
      summary: Get one device
      tags: [devices]
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Device" }
  /device-status:
    get:
      summary: Live health snapshot (battery, connectivity, signal quality)
      tags: [devices]
      parameters:
        - $ref: "#/components/parameters/DeviceIdQuery"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeviceStatus" }
...
  /devices/{device_id}/ai/insight:
    get:
      summary: >
        Phase 10 -- a Gemini-generated, plain-language summary of the
        device's recent readings and active alerts. Requires auth. 503 if
        the backend has no GEMINI_API_KEY configured; 404 if the device is
        unknown or has no measurement history yet.
      tags: [ai]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AIInsight" }
        "404": { description: Device not found, or no measurement history yet }
        "502": { description: The Gemini API call failed }
        "503": { description: GEMINI_API_KEY is not configured on the backend }
  /devices/{device_id}/ai/chat:
    post:
      summary: >
        Phase 10 -- a follow-up question about the device's data, answered
        by Gemini grounded in the same context as /ai/insight. Requires
        auth.
      tags: [ai]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: { type: string, minLength: 1 }
                history:
                  type: array
                  description: Prior turns, oldest first
                  items: { $ref: "#/components/schemas/AIChatMessage" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AIChatReply" }
        "404": { description: Device not found, or no measurement history yet }
        "502": { description: The Gemini API call failed }
        "503": { description: GEMINI_API_KEY is not configured on the backend }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    DeviceId:
      name: device_id
      in: path
      required: true
      schema: { type: string }
    DeviceIdQuery:
      name: device_id
      in: query
      required: true
      schema: { type: string }
  schemas:
    Device:
      type: object
      required: [device_id, name, firmware_version]
      properties:
        device_id: { type: string, example: SB-001 }
        name: { type: string, example: "Prototype #1" }
        firmware_version: { type: string, example: "0.1.0" }
        channels:
          type: array
          items: { type: string }
        status: { type: string, enum: [online, offline, unknown] }
        last_seen: { type: string, format: date-time }
    DeviceStatus:
      type: object
      required: [connected]
      properties:
        connected: { type: boolean }
        battery: { type: integer, minimum: 0, maximum: 100 }
        signal_quality: { type: number, minimum: 0, maximum: 1 }
        last_error: { type: string, nullable: true }
...
    AIInsight:
      type: object
      required: [device_id, summary, model, generated_at]
      properties:
        device_id: { type: string }
        summary: { type: string, description: "Plain-language, Gemini-generated -- not a diagnosis" }
        model: { type: string, example: gemini-3.6-flash }
        generated_at: { type: string, format: date-time }
    AIChatMessage:
      type: object
      required: [role, text]
      properties:
        role: { type: string, enum: [user, model] }
        text: { type: string }
    AIChatReply:
      type: object
      required: [device_id, reply, model]
      properties:
        device_id: { type: string }
        reply: { type: string }
        model: { type: string, example: gemini-3.6-flash }